rename template name
new-event -> event
