+++ workers/email/src/templates/emails/event/text.twig
