+++ workers/slack/src/templates/event.ts
