+++ workers/slack/src/templates/index.ts
@@ -1,10 +1,10 @@
 /**
  * This file exports all available templates for Slack notification
  */
-import NewEventTpl from './new-event';
+import EventTpl from './event';
 import SeveralEventsTpl from './several-events';
 
 export default {
-  NewEventTpl,
+  EventTpl,
   SeveralEventsTpl,
 };
