+++ workers/telegram/src/templates/event.ts
