+++ workers/telegram/src/templates/index.ts
@@ -1,7 +1,7 @@
-import NewEventTpl from './new-event';
+import EventTpl from './event';
 import SeveralEventsTpl from './several-events';
 
 export default {
-  NewEventTpl,
+  EventTpl,
   SeveralEventsTpl,
 };
